separate Runner from LocalRunner
